test(shared): consolidate duplicate suite pairs
