+++ packages/shared/src/apps/overlay-app-registry.test.ts
@@ -2,7 +2,7 @@
  * Unit tests for full-screen overlay app registry and platform availability filters.
  */
 
-import { describe, expect, it } from "vitest";
+import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import type { OverlayApp } from "./overlay-app-api.js";
 import {
   getAllOverlayApps,
@@ -13,6 +13,20 @@ import {
   overlayAppToRegistryInfo,
   registerOverlayApp,
 } from "./overlay-app-registry.js";
+
+const store = new Map<string, unknown>();
+
+vi.mock("../registry-host.js", () => ({
+  getUiRegistryStore: (key: string, factory: () => unknown) => {
+    if (!store.has(key)) store.set(key, factory());
+    return store.get(key);
+  },
+}));
+
+vi.mock("../platform/aosp-user-agent.js", () => ({
+  userAgentHasElizaOSMarker: (userAgent: string) =>
+    userAgent.includes("ElizaOS/"),
+}));
 
 function makeOverlayApp(overrides: Partial<OverlayApp> = {}): OverlayApp {
   return {
@@ -26,6 +40,9 @@ function makeOverlayApp(overrides: Partial<OverlayApp> = {}): OverlayApp {
   };
 }
 
+beforeEach(() => store.clear());
+afterEach(() => store.clear());
+
 describe("overlay app registry", () => {
   it("registers and retrieves overlay apps by name", () => {
     const app = makeOverlayApp({ name: "app-one", displayName: "App One" });
@@ -36,6 +53,14 @@ describe("overlay app registry", () => {
     expect(getOverlayApp("app-one")).toBe(app);
     expect(getOverlayApp("nonexistent-app")).toBeUndefined();
     expect(getAllOverlayApps()).toContain(app);
+  });
+
+  it("overwrites an existing registration with the same name", () => {
+    registerOverlayApp(makeOverlayApp({ name: "same", displayName: "Old" }));
+    registerOverlayApp(makeOverlayApp({ name: "same", displayName: "New" }));
+
+    expect(getAllOverlayApps()).toHaveLength(1);
+    expect(getOverlayApp("same")?.displayName).toBe("New");
   });
 
   it("filters out androidOnly apps on non-AOSP platforms", () => {
@@ -94,6 +119,23 @@ describe("overlay app registry", () => {
     expect(isAospAndroid({ platform: "ios", aospAndroid: true })).toBe(false);
   });
 
+  it("supports plain platform strings and detects the AOSP user-agent marker", () => {
+    registerOverlayApp(
+      makeOverlayApp({ name: "aosp-only", androidOnly: true }),
+    );
+
+    expect(getAvailableOverlayApps("android")).toHaveLength(0);
+    expect(
+      isAospAndroid({
+        platform: "android",
+        userAgent: "Mozilla ElizaOS/1.2.3",
+      }),
+    ).toBe(true);
+    expect(
+      isAospAndroid({ platform: "android", userAgent: "Mozilla plain" }),
+    ).toBe(false);
+  });
+
   it("converts OverlayApp to RegistryAppInfo format", () => {
     const app = makeOverlayApp({
       name: "converter-test-app",
@@ -127,5 +169,9 @@ describe("overlay app registry", () => {
         v2Version: null,
       },
     });
+  });
+
+  it("defaults a missing hero image to null", () => {
+    expect(overlayAppToRegistryInfo(makeOverlayApp()).heroImage).toBeNull();
   });
 });
